harness: extract TestCompiler and auto_discovery into harness package
Reduces the shim from 3354 lines to 123 lines.

- tests/harness/test_compiler.py: TestCompiler class + UNIT_ONLY_TESTS/FLAGS
- tests/harness/auto_discovery.py: _discover_and_add_{auto,fpc_rtl}_tests + _add_pp_pas_bootstrap_test + classification sets (KGPC_ONLY_TESTS, FPC_RTL_ONLY_TESTS, FPC_RTL_IMPLICIT_UNIT_TESTS)
- tests/do_not_run_me_directly_but_through_meson.py: thin shim (~123 lines) that imports TestCompiler + auto_discovery (side-effect populates TestCompiler) then provides _load_suite() and main()

Compiler tests: 1035 subtests passed, 0 failures.
FPC RTL tests: 234 ok, 3 pre-existing failures (pp_pas_bootstrap, tdatetime_basics, adfgvxcipher timeouts).
